Move LoginViewModel to LoginActivity inner class.  Stop passing ViewModel.  Add a view model factory var on SDKMangager so LoginActivity can be customized without overriding.
