+++ libs/SalesforceSDK/src/com/salesforce/androidsdk/ui/LoginActivity.kt
@@ -56,6 +56,9 @@ import android.view.Display.FLAG_SECURE
 import android.view.KeyEvent
 import android.view.KeyEvent.KEYCODE_BACK
 import android.view.View
+import android.webkit.WebResourceRequest
+import android.webkit.WebView
+import android.webkit.WebViewClient
 import android.widget.Button
 import android.widget.Toast.LENGTH_LONG
 import android.widget.Toast.LENGTH_SHORT
@@ -82,6 +85,7 @@ import androidx.biometric.BiometricPrompt.AuthenticationCallback
 import androidx.biometric.BiometricPrompt.AuthenticationResult
 import androidx.biometric.BiometricPrompt.PromptInfo
 import androidx.browser.customtabs.CustomTabsIntent
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.ComposeView
 import androidx.core.content.ContextCompat.getMainExecutor
 import androidx.fragment.app.FragmentActivity
@@ -99,17 +103,14 @@ import com.salesforce.androidsdk.accounts.UserAccountManager.USER_SWITCH_TYPE_LO
 import com.salesforce.androidsdk.analytics.SalesforceAnalyticsManager
 import com.salesforce.androidsdk.app.SalesforceSDKManager
 import com.salesforce.androidsdk.auth.LoginViewModel
-import com.salesforce.androidsdk.auth.LoginWebviewClient
 import com.salesforce.androidsdk.auth.OAuth2.OAuthFailedException
+import com.salesforce.androidsdk.auth.OAuth2.TokenEndpointResponse
 import com.salesforce.androidsdk.auth.idp.interfaces.SPManager.Status
 import com.salesforce.androidsdk.auth.idp.interfaces.SPManager.StatusUpdateCallback
 import com.salesforce.androidsdk.config.RuntimeConfig.ConfigKey.ManagedAppCertAlias
 import com.salesforce.androidsdk.config.RuntimeConfig.ConfigKey.RequireCertAuth
 import com.salesforce.androidsdk.config.RuntimeConfig.getRuntimeConfig
-import com.salesforce.androidsdk.rest.ClientManager.LoginOptions
-import com.salesforce.androidsdk.rest.ClientManager.LoginOptions.fromBundleWithSafeLoginUrl
 import com.salesforce.androidsdk.security.BiometricAuthenticationManager
-import com.salesforce.androidsdk.security.BiometricAuthenticationManager.Companion.SHOW_BIOMETRIC
 import com.salesforce.androidsdk.ui.OAuthWebviewHelper.Companion.AUTHENTICATION_FAILED_INTENT
 import com.salesforce.androidsdk.ui.OAuthWebviewHelper.Companion.HTTP_ERROR_RESPONSE_CODE_INTENT
 import com.salesforce.androidsdk.ui.OAuthWebviewHelper.Companion.RESPONSE_ERROR_DESCRIPTION_INTENT
@@ -117,10 +118,12 @@ import com.salesforce.androidsdk.ui.OAuthWebviewHelper.Companion.RESPONSE_ERROR_
 import com.salesforce.androidsdk.ui.components.LoginView
 import com.salesforce.androidsdk.ui.theme.LoginWebviewTheme
 import com.salesforce.androidsdk.util.EventsObservable
+import com.salesforce.androidsdk.util.EventsObservable.EventType.AuthWebViewPageFinished
 import com.salesforce.androidsdk.util.EventsObservable.EventType.LoginActivityCreateComplete
 import com.salesforce.androidsdk.util.SalesforceSDKLogger.d
 import com.salesforce.androidsdk.util.SalesforceSDKLogger.e
 import com.salesforce.androidsdk.util.SalesforceSDKLogger.w
+import com.salesforce.androidsdk.util.UriFragmentParser
 import com.salesforce.androidsdk.util.UriFragmentParser.parse
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers.IO
@@ -132,7 +135,9 @@ import java.security.cert.X509Certificate
 
 
 open class LoginActivity: FragmentActivity() {
-    protected open val viewModel: LoginViewModel by viewModels { LoginViewModel.Factory }
+    protected open val viewModel: LoginViewModel
+        by viewModels { SalesforceSDKManager.getInstance().loginViewModelFactory }
+    protected open val webviewClient = LoginWebviewClient()
 
     private var wasBackgrounded = false
     private var accountAuthenticatorResponse: AccountAuthenticatorResponse? = null
@@ -178,7 +183,7 @@ open class LoginActivity: FragmentActivity() {
             ComposeView(this).apply {
                 setContent {
                     LoginWebviewTheme {
-                        LoginView(webviewClient = LoginWebviewClient(viewModel, ::onAuthFlowError, ::onAuthFlowSuccess))
+                        LoginView(webviewClient)
                     }
                 }
             }
@@ -656,8 +661,19 @@ open class LoginActivity: FragmentActivity() {
                 .build()
         }
 
-    // TODO: move this up?
     open fun onBioAuthClick(view: View?) = presentBiometric()
+
+    /**
+     * Called when the IDP login button is clicked.
+     *
+     * @param v IDP login button
+     */
+    open fun onIDPLoginClick(v: View?) {
+        SalesforceSDKManager.getInstance().spManager?.kickOffSPInitiatedLoginFlow(
+            this,
+            SPStatusCallback()
+        )
+    }
 
     private fun loadLoginPageInCustomTab(loginUrl: String, customTabLauncher: ActivityResultLauncher<Intent>) {
         val customTabsIntent = CustomTabsIntent.Builder().apply {
@@ -734,21 +750,99 @@ open class LoginActivity: FragmentActivity() {
     // endregion
 
     /**
-     * Reloads the authorization page in the web view.  Also, updates the window
-     * title so it's easier to identify the login system.
+     * LoginWebviewClient is an inner class of LoginActivity because it makes extensive use of the LoginViewModel,
+     * which is only available to Activity classes (and composable functions).
      */
-//    private fun loadLoginPage() = loginOptions.let { loginOptions ->
-//        when {
-//            isEmpty(loginOptions.jwt) -> {
-//                loginOptions.loginUrl = viewModel.loginUrl
-//                doLoadPage()
-//            }
-//
-//            else -> CoroutineScope(IO).launch {
-//                SwapJWTForAccessTokenTask().execute(loginOptions)
-//            }
-//        }
-//    }
+    open inner class LoginWebviewClient: WebViewClient() {
+        override fun shouldOverrideUrlLoading(view: WebView, request: WebResourceRequest): Boolean {
+            // Check if user entered a custom domain
+            val customDomainPatternMatch = SalesforceSDKManager.getInstance()
+                .customDomainInferencePattern?.matcher(request.url.toString())?.find() ?: false
+            val loginContainsHost = request.url.host?.let { viewModel.selectedServer.value?.contains(it) } ?: false
+            if (customDomainPatternMatch && !loginContainsHost) {
+                runCatching {
+                    val baseUrl = "https://${request.url.host}"
+                    val serverManager = SalesforceSDKManager.getInstance().loginServerManager
+
+                    // Check if the URL is already in the server list
+                    when (val loginServer = serverManager.getLoginServerFromURL(baseUrl)) {
+                        null ->
+                            // Add also sets as selected
+                            serverManager.addCustomLoginServer("Custom Domain", baseUrl)
+                        else ->
+                            serverManager.selectedLoginServer = loginServer
+                    }
+                }.onFailure { throwable ->
+                    e(TAG, "Unable to retrieve auth config.", throwable)
+                }
+            }
+
+            val formattedUrl = request.url.toString().replace("///", "/").lowercase()
+            val callbackUrl = viewModel.bootConfig.oauthRedirectURI.replace("///", "/").lowercase()
+            val authFlowFinished = formattedUrl.startsWith(callbackUrl)
+
+            if (authFlowFinished) {
+                val params = UriFragmentParser.parse(request.url)
+                val error = params["error"]
+                // Did we fail?
+                when {
+                    error != null -> onAuthFlowError(
+                        error,
+                        params["error_description"],
+                        null
+                    )
+
+                    else -> {
+                        // Determine if presence of override parameters require the user agent flow.
+                        val overrideWithUserAgentFlow = viewModel.isUsingFrontDoorBridge && viewModel.frontDoorBridgeCodeVerifier == null
+                        when {
+                            SalesforceSDKManager.getInstance().useWebServerAuthentication && !overrideWithUserAgentFlow ->
+                                viewModel.onWebServerFlowComplete(params["code"], ::onAuthFlowError, ::onAuthFlowSuccess)
+
+                            else ->
+                                viewModel.onAuthFlowComplete(TokenEndpointResponse(params), ::onAuthFlowError, ::onAuthFlowSuccess)
+                        }
+                    }
+                }
+            }
+
+            return authFlowFinished
+        }
+
+        override fun onPageFinished(view: WebView?, url: String?) {
+            view?.evaluateJavascript(backgroundColorJavascript) { result ->
+                viewModel.loading.value = false
+                if (url == ABOUT_BLANK) {
+                    viewModel.dynamicBackgroundColor.value = Color.White
+                    return@evaluateJavascript
+                }
+
+                viewModel.dynamicBackgroundColor.value = validateAndExtractBackgroundColor(result) ?: return@evaluateJavascript
+            }
+
+            // Remove the native login buttons (biometric, IDP) once on the allow/deny screen
+            if (url?.contains(ALLOW_SCREEN_INDICATOR) == true) {
+                // TODO: hide buttons via viewModel
+            }
+            EventsObservable.get().notifyEvent(AuthWebViewPageFinished, url)
+
+            super.onPageFinished(view, url)
+        }
+
+        private fun validateAndExtractBackgroundColor(javaScriptResult: String): Color? {
+            // This parses the expected "rgb(x, x, x)" string.
+            val rgbTextPattern = "rgb\\((\\d{1,3}), (\\d{1,3}), (\\d{1,3})\\)".toRegex()
+            val rgbMatch = rgbTextPattern.find(javaScriptResult)
+
+            // groupValues[0] is the entire match.  [1] is red, [2] is green, [3] is green.
+            rgbMatch?.groupValues?.get(3) ?: return null
+            val red = rgbMatch.groupValues[1].toIntOrNull() ?: return null
+            val green = rgbMatch.groupValues[2].toIntOrNull() ?: return null
+            val blue = rgbMatch.groupValues[3].toIntOrNull() ?: return null
+
+            return Color(red, green, blue)
+        }
+    }
 
 
     companion object {
@@ -759,6 +853,14 @@ open class LoginActivity: FragmentActivity() {
         private const val PROMPT_LOGIN = "&prompt=login"
 
         // endregion
+
+        // region LoginWebviewClient Constants
+        private const val ABOUT_BLANK = "about:blank"
+        private const val ALLOW_SCREEN_INDICATOR = "frontdoor.jsp"
+        private const val backgroundColorJavascript =
+            "(function() { return window.getComputedStyle(document.body, null).getPropertyValue('background-color'); })();"
+        // endregion
+
         // region QR Code Login Via Salesforce Identity API UI Bridge Public Implementation
 
         /**
