+++ libs/SalesforceSDK/src/com/salesforce/androidsdk/ui/components/LoginWebview.kt
@@ -3,6 +3,7 @@ package com.salesforce.androidsdk.ui.components
 import android.annotation.SuppressLint
 import android.view.ViewGroup
 import android.webkit.WebView
+import android.webkit.WebViewClient
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
@@ -12,17 +13,19 @@ import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.viewinterop.AndroidView
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.salesforce.androidsdk.app.SalesforceSDKManager
 import com.salesforce.androidsdk.auth.LoginViewModel
-import com.salesforce.androidsdk.auth.LoginWebviewClient
 
 @SuppressLint("SetJavaScriptEnabled")
 @Composable
 fun LoginWebview(
     paddingValues: PaddingValues,
-    webviewClient: LoginWebviewClient,
-    viewModel: LoginViewModel = viewModel(factory = LoginViewModel.Factory),
+    webviewClient: WebViewClient,
+    viewModelFactory: ViewModelProvider.Factory = SalesforceSDKManager.getInstance().loginViewModelFactory,
 ) {
+    val viewModel: LoginViewModel = viewModel(factory = viewModelFactory)
     val loginUrl: String = viewModel.loginUrl.observeAsState().value ?: ""
 
     AndroidView(
